lib/posix-pipe: Split pipe functionality into separate library
This library provides the pipe/pipe2 system calls and remove the vfscore
dependency.
