+++ lib/Makefile.uk
@@ -16,6 +16,7 @@ $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-mmap))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-process))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-socket))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-sysinfo))
+$(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-pipe))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-futex))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/posix-user))
 $(eval $(call _import_lib,$(CONFIG_UK_BASE)/lib/ramfs))
