+++ lib/posix-pipe/Makefile.uk
@@ -1,0 +1,6 @@
+$(eval $(call addlib_s,libposix-pipe,$(CONFIG_LIBPOSIX_PIPE)))
+
+LIBPOSIX-PIPE_SRCS-y += $(LIBPOSIX-PIPE_BASE)/pipe.c
+
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_PIPE) += pipe2-2
+UK_PROVIDED_SYSCALLS-$(CONFIG_LIBPOSIX_PIPE) += pipe-1
