+++ lib/posix-pipe/pipe.c
@@ -1,0 +1,623 @@
+/* SPDX-License-Identifier: BSD-3-Clause */
+/*
+ * Authors: Bogdan-George Lascu <lascu.bogdan96@gmail.com>
+ *          Costin Lupu <costin.lupu@cs.pub.ro>
+ *
+ * Copyright (c) 2019, University Politehnica of Bucharest. All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the distribution.
+ * 3. Neither the name of the copyright holder nor the names of its
+ *    contributors may be used to endorse or promote products derived from
+ *    this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include <uk/config.h>
+#include <stdlib.h>
+#include <stdio.h>
+#include <string.h>
+#include <uk/fdtab/fd.h>
+#include <uk/fdtab/eventpoll.h>
+#include <uk/wait.h>
+#include <uk/syscall.h>
+#include <sys/ioctl.h>
+
+/* We use the default size in Linux kernel */
+#define PIPE_MAX_SIZE	(1 << CONFIG_LIBPOSIX_PIPE_SIZE_ORDER)
+
+#define to_pipe(fp) __containerof((fp), struct pipe, file)
+
+struct pipe_buf {
+	/* The buffer */
+	char *data;
+	/* The buffer capacity, always a power of 2 */
+	unsigned long capacity;
+	/* Producer index */
+	unsigned long prod;
+	/* Consumer index */
+	unsigned long cons;
+
+	/* Read lock */
+	struct uk_mutex rdlock;
+	/* Write lock */
+	struct uk_mutex wrlock;
+
+	/* Readers queue */
+	struct uk_waitq rdwq;
+	/* Writers queue */
+	struct uk_waitq wrwq;
+};
+
+#define PIPE_BUF_IDX(buf, n)    ((n) & ((buf)->capacity - 1))
+#define PIPE_BUF_PROD_IDX(buf)  PIPE_BUF_IDX((buf), (buf)->prod)
+#define PIPE_BUF_CONS_IDX(buf)  PIPE_BUF_IDX((buf), (buf)->cons)
+
+struct pipe_file {
+	/* Pipe buffer */
+	struct pipe_buf *buf;
+	/* Write reference count */
+	int w_refcount;
+	/* Read reference count */
+	int r_refcount;
+	/* Flags */
+	int flags;
+
+	/* Eventpoll lock */
+	struct uk_mutex evp_lock;
+	/* List of registered eventpolls */
+	struct uk_list_head evp_list;
+};
+
+struct pipe {
+	struct fdtab_file file;
+	struct pipe_file *pipe_file;
+};
+
+static struct pipe_buf *pipe_buf_alloc(int capacity)
+{
+	struct pipe_buf *pipe_buf;
+
+	UK_ASSERT(POWER_OF_2(capacity));
+
+	pipe_buf = malloc(sizeof(*pipe_buf));
+	if (!pipe_buf)
+		return NULL;
+
+	pipe_buf->data = malloc(capacity);
+	if (!pipe_buf->data) {
+		free(pipe_buf);
+		return NULL;
+	}
+
+	pipe_buf->capacity = capacity;
+	pipe_buf->cons = 0;
+	pipe_buf->prod = 0;
+	uk_mutex_init(&pipe_buf->rdlock);
+	uk_mutex_init(&pipe_buf->wrlock);
+	uk_waitq_init(&pipe_buf->rdwq);
+	uk_waitq_init(&pipe_buf->wrwq);
+
+	return pipe_buf;
+}
+
+void pipe_buf_free(struct pipe_buf *pipe_buf)
+{
+	free(pipe_buf->data);
+	free(pipe_buf);
+}
+
+static unsigned long pipe_buf_get_available(const struct pipe_buf *pipe_buf)
+{
+	return pipe_buf->prod - pipe_buf->cons;
+}
+
+static unsigned long pipe_buf_get_free_space(struct pipe_buf *pipe_buf)
+{
+	return pipe_buf->capacity - pipe_buf_get_available(pipe_buf);
+}
+
+static int pipe_buf_can_write(struct pipe_buf *pipe_buf)
+{
+	return pipe_buf_get_free_space(pipe_buf) > 0;
+}
+
+static int pipe_buf_can_read(struct pipe_buf *pipe_buf)
+{
+	return pipe_buf_get_available(pipe_buf) > 0;
+}
+
+static int pipe_file_can_read(struct pipe_file *pipe_file)
+{
+	return pipe_buf_can_read(pipe_file->buf) || !(pipe_file->w_refcount);
+}
+
+static unsigned long pipe_buf_write(struct pipe_buf *pipe_buf,
+				    struct iovec *iovec, size_t iovec_off)
+{
+	unsigned long prod_idx, to_write;
+	void *iovec_data = iovec->iov_base + iovec_off;
+	size_t iov_len = iovec->iov_len - iovec_off;
+
+	prod_idx = PIPE_BUF_PROD_IDX(pipe_buf);
+	to_write = MIN(pipe_buf_get_free_space(pipe_buf), iov_len);
+	if (to_write == 0)
+		goto out;
+
+	/* Copy in one piece */
+	if (prod_idx + to_write <= pipe_buf->capacity)
+		memcpy(pipe_buf->data + prod_idx, iovec_data, to_write);
+
+	else {
+		int first_copy_bytes, second_copy_bytes;
+
+		/* Copy the first part */
+		first_copy_bytes = pipe_buf->capacity - prod_idx;
+		memcpy(pipe_buf->data + prod_idx,
+		       iovec_data,
+		       first_copy_bytes);
+
+		/* Copy the second part */
+		second_copy_bytes = prod_idx + to_write - pipe_buf->capacity;
+		memcpy(pipe_buf->data,
+		       iovec_data + first_copy_bytes,
+		       second_copy_bytes);
+	}
+
+	/* Update producer */
+	pipe_buf->prod += to_write;
+
+out:
+	return to_write;
+}
+
+static unsigned long pipe_buf_read(struct pipe_buf *pipe_buf,
+				   struct iovec *iovec, size_t iovec_off)
+{
+	unsigned long cons_idx, to_read;
+	void *iovec_data = iovec->iov_base + iovec_off;
+	size_t iov_len = iovec->iov_len - iovec_off;
+
+	cons_idx = PIPE_BUF_CONS_IDX(pipe_buf);
+	to_read = MIN(pipe_buf_get_available(pipe_buf), iov_len);
+	if (to_read == 0)
+		goto out;
+
+	/* Copy in one piece */
+	if (cons_idx + to_read <= pipe_buf->capacity)
+		memcpy(iovec_data, pipe_buf->data + cons_idx, to_read);
+
+	else {
+		int first_copy_bytes;
+		int second_copy_bytes;
+
+		/* Copy the first part */
+		first_copy_bytes = pipe_buf->capacity - pipe_buf->cons;
+		memcpy(iovec_data,
+		       pipe_buf->data + cons_idx,
+		       first_copy_bytes);
+
+		/* Copy the second part */
+		second_copy_bytes = cons_idx + to_read - pipe_buf->capacity;
+		memcpy(iovec_data + first_copy_bytes,
+		       pipe_buf->data,
+		       second_copy_bytes);
+	}
+
+	/* Update consumer */
+	pipe_buf->cons += to_read;
+
+out:
+	return to_read;
+}
+
+struct pipe_file *pipe_file_alloc(int capacity, int flags)
+{
+	struct pipe_file *pipe_file;
+
+	pipe_file = malloc(sizeof(*pipe_file));
+	if (!pipe_file)
+		return NULL;
+
+	pipe_file->buf = pipe_buf_alloc(capacity);
+	if (!pipe_file->buf) {
+		free(pipe_file);
+		return NULL;
+	}
+
+	pipe_file->w_refcount = 1;
+	pipe_file->r_refcount = 1;
+	pipe_file->flags = flags;
+
+	uk_mutex_init(&pipe_file->evp_lock);
+	UK_INIT_LIST_HEAD(&pipe_file->evp_list);
+
+	return pipe_file;
+}
+
+void pipe_file_free(struct pipe_file *pipe_file)
+{
+	pipe_buf_free(pipe_file->buf);
+	free(pipe_file);
+}
+
+static void
+pipe_file_event(struct pipe_file *pipe_file, unsigned int event)
+{
+	struct eventpoll_cb *ecb;
+	struct uk_list_head *itr;
+
+	UK_ASSERT(pipe_file);
+
+	uk_mutex_lock(&pipe_file->evp_lock);
+	uk_list_for_each(itr, &pipe_file->evp_list) {
+		ecb = uk_list_entry(itr, struct eventpoll_cb, cb_link);
+
+		UK_ASSERT(ecb->unregister);
+
+		eventpoll_signal(ecb, event);
+	}
+	uk_mutex_unlock(&pipe_file->evp_lock);
+}
+
+static void
+pipe_file_unregister_eventpoll(struct eventpoll_cb *ecb)
+{
+	struct pipe_file *pipe_file;
+
+	UK_ASSERT(ecb);
+	UK_ASSERT(ecb->data);
+	pipe_file = (struct pipe_file *)ecb->data;
+
+	uk_mutex_lock(&pipe_file->evp_lock);
+	UK_ASSERT(!uk_list_empty(&ecb->cb_link));
+	uk_list_del(&ecb->cb_link);
+
+	ecb->data = NULL;
+	ecb->unregister = NULL;
+	uk_mutex_unlock(&pipe_file->evp_lock);
+}
+
+static int pipe_write(struct fdtab_file *fp, struct uio *buf,
+		      int ioflag __unused)
+{
+	struct pipe_file *pipe_file = to_pipe(fp)->pipe_file;
+	struct pipe_buf *pipe_buf = pipe_file->buf;
+	bool nonblocking = false; /* TODO handle nonblocking */
+	bool data_available = true;
+	int uio_idx = 0;
+
+	if (!pipe_file->r_refcount) {
+		/* TODO before returning the error, send a SIGPIPE signal */
+		return -EPIPE;
+	}
+
+	uk_mutex_lock(&pipe_buf->wrlock);
+	while (data_available && uio_idx < buf->uio_iovcnt) {
+		struct iovec *iovec = &buf->uio_iov[uio_idx];
+		unsigned long off = 0;
+
+		while (off < iovec->iov_len) {
+			unsigned long written_bytes;
+
+			written_bytes = pipe_buf_write(pipe_buf, iovec, off);
+			if (written_bytes == 0) {
+				/* No data */
+				if (nonblocking) {
+					data_available = false;
+					break;
+
+				} else {
+					/* Wait until data available */
+					while (!pipe_buf_can_write(pipe_buf)) {
+						uk_mutex_unlock(
+						    &pipe_buf->wrlock);
+						uk_waitq_wait_event(
+						    &pipe_buf->wrwq,
+						    pipe_buf_can_write(
+							pipe_buf));
+						uk_mutex_lock(
+						    &pipe_buf->wrlock);
+					}
+				}
+
+			} else {
+				/* Update bytes written_bytes. */
+				buf->uio_resid -= written_bytes;
+
+				off += written_bytes;
+
+				/* wake some readers */
+				uk_waitq_wake_up(&pipe_buf->rdwq);
+				pipe_file_event(pipe_file,
+						EPOLLIN | EPOLLRDNORM);
+			}
+		}
+
+		uio_idx++;
+	}
+	uk_mutex_unlock(&pipe_buf->wrlock);
+
+	return 0;
+}
+
+static int pipe_read(struct fdtab_file *fp, struct uio *buf,
+		     int ioflag __unused)
+{
+	struct pipe_file *pipe_file = to_pipe(fp)->pipe_file;
+	struct pipe_buf *pipe_buf = pipe_file->buf;
+	bool nonblocking = (fp->f_flags & O_NONBLOCK);
+	bool data_available = true;
+	int uio_idx = 0;
+
+	uk_mutex_lock(&pipe_buf->rdlock);
+	if (nonblocking && !pipe_buf_can_read(pipe_buf)) {
+		uk_mutex_unlock(&pipe_buf->rdlock);
+		return EAGAIN;
+	}
+
+	while (data_available && uio_idx < buf->uio_iovcnt) {
+		struct iovec *iovec = &buf->uio_iov[uio_idx];
+		unsigned long off = 0;
+
+		while (off < iovec->iov_len) {
+			unsigned long read_bytes;
+
+			read_bytes = pipe_buf_read(pipe_buf, iovec, off);
+			if (read_bytes == 0) {
+				/* No data */
+				if (nonblocking) {
+					data_available = false;
+					break;
+
+				} else {
+					/* Wait until data available */
+					while (!pipe_buf_can_read(pipe_buf)) {
+						uk_mutex_unlock(
+						    &pipe_buf->rdlock);
+						uk_waitq_wait_event(
+						    &pipe_buf->rdwq,
+						    pipe_file_can_read(
+							pipe_file));
+
+						if (!pipe_file->w_refcount)
+							return 0;
+
+						uk_mutex_lock(
+						    &pipe_buf->rdlock);
+					}
+				}
+
+			} else {
+				/* Update bytes read */
+				buf->uio_resid -= read_bytes;
+
+				off += read_bytes;
+
+				/* wake some writers */
+				uk_waitq_wake_up(&pipe_buf->wrwq);
+				pipe_file_event(pipe_file,
+						EPOLLOUT | EPOLLWRNORM);
+
+				break;
+			}
+		}
+
+		uio_idx++;
+	}
+	uk_mutex_unlock(&pipe_buf->rdlock);
+
+	return 0;
+}
+
+static int pipe_free(struct fdtab_file *fp)
+{
+	struct pipe_file *pipe_file = to_pipe(fp)->pipe_file;
+
+	if (fp->f_flags & UK_FREAD)
+		pipe_file->r_refcount--;
+
+	if (fp->f_flags & UK_FWRITE)
+		pipe_file->w_refcount--;
+
+	if (!pipe_file->r_refcount && !pipe_file->w_refcount)
+		pipe_file_free(pipe_file);
+
+	return 0;
+}
+
+static int pipe_ioctl(struct fdtab_file *fp, unsigned long com, void *data)
+{
+	struct pipe_file *pipe_file = to_pipe(fp)->pipe_file;
+	struct pipe_buf *pipe_buf = pipe_file->buf;
+
+	switch (com) {
+	case FIONREAD:
+		uk_mutex_lock(&pipe_buf->rdlock);
+		*((int *) data) = pipe_buf_get_available(pipe_buf);
+		uk_mutex_unlock(&pipe_buf->rdlock);
+		return 0;
+	default:
+		return -EINVAL;
+	}
+}
+
+static unsigned int
+get_pipe_file_events(struct pipe_file *pipe_file)
+{
+	struct pipe_buf *pipe_buf = pipe_file->buf;
+	unsigned int events = 0;
+
+	UK_ASSERT(pipe_file);
+
+	uk_mutex_lock(&pipe_buf->rdlock);
+	if (pipe_buf_can_read(pipe_buf))
+		events |= EPOLLIN | EPOLLRDNORM;
+	uk_mutex_unlock(&pipe_buf->rdlock);
+
+	uk_mutex_lock(&pipe_buf->wrlock);
+	if (pipe_buf_can_write(pipe_buf))
+		events |= EPOLLOUT | EPOLLWRNORM;
+	uk_mutex_unlock(&pipe_buf->wrlock);
+
+	return events;
+}
+
+static int
+pipe_poll(struct fdtab_file *fp, unsigned int *revents,
+	  struct eventpoll_cb *ecb)
+{
+	struct pipe_file *pipe_file = to_pipe(fp)->pipe_file;
+
+	UK_ASSERT(revents);
+
+	*revents = get_pipe_file_events(pipe_file);
+
+	uk_mutex_lock(&pipe_file->evp_lock);
+	if (!ecb->unregister) {
+		UK_ASSERT(uk_list_empty(&ecb->cb_link));
+		UK_ASSERT(!ecb->data);
+		/* This is the first time we see this cb. Add it to the
+		 * eventpoll list and set the unregister callback so
+		 * we remove it when the eventpoll is freed.
+		 */
+		uk_list_add_tail(&ecb->cb_link, &pipe_file->evp_list);
+
+		ecb->data = pipe_file;
+		ecb->unregister = pipe_file_unregister_eventpoll;
+	}
+	uk_mutex_unlock(&pipe_file->evp_lock);
+
+	return 0;
+}
+
+static struct fdops pipe_vnops = {
+	.fdop_free	= pipe_free,
+	.fdop_read	= pipe_read,
+	.fdop_write	= pipe_write,
+	.fdop_ioctl	= pipe_ioctl,
+	.fdop_poll	= pipe_poll,
+};
+
+static int pipe_fd_alloc(struct fdtab_table *tab, struct pipe_file *pipe_file,
+			 int flags)
+{
+	int ret;
+	int fd;
+	struct pipe *pipe = NULL;
+	struct fdtab_file *fp;
+
+	/* Reserve file descriptor number */
+	fd = fdtab_alloc_fd(tab);
+	if (fd < 0) {
+		ret = -ENFILE;
+		goto ERR_EXIT;
+	}
+
+	/* Allocate file, dentry, and vnode */
+	pipe = calloc(1, sizeof(*pipe));
+	if (!pipe) {
+		ret = -ENOMEM;
+		goto ERR_ALLOC_FD;
+	}
+
+	fp = &pipe->file;
+	fdtab_file_init(fp);
+
+	/* Fill out necessary fields. */
+	fp->fd = fd;
+	fp->f_flags = flags;
+	fp->f_count = 1;
+	fp->f_op = &pipe_vnops;
+	pipe->pipe_file = pipe_file;
+
+	/* Assign the file descriptors to the corresponding vfs_file. */
+	ret = fdtab_install_fd(tab, fd, fp);
+	if (ret)
+		goto ERR_MALLOC_PIPE;
+
+	return fd;
+
+ERR_MALLOC_PIPE:
+	free(pipe);
+ERR_ALLOC_FD:
+	fdtab_put_fd(tab, fd);
+ERR_EXIT:
+	UK_ASSERT(ret < 0);
+	return ret;
+}
+
+UK_SYSCALL_R_DEFINE(int, pipe, int*, pipefd)
+{
+	int ret = 0;
+	int r_fd, w_fd;
+	struct pipe_file *pipe_file;
+	struct fdtab_table *tab = fdtab_get_active();
+
+	/* Allocate pipe internal structure. */
+	pipe_file = pipe_file_alloc(PIPE_MAX_SIZE, 0);
+	if (!pipe_file) {
+		ret = -ENOMEM;
+		goto ERR_EXIT;
+	}
+
+	r_fd = pipe_fd_alloc(tab, pipe_file, UK_FREAD);
+	if (r_fd < 0)
+		goto ERR_VFS_INSTALL;
+
+	w_fd = pipe_fd_alloc(tab, pipe_file, UK_FWRITE);
+	if (w_fd < 0)
+		goto ERR_W_FD;
+
+	/* Fill pipefd fields. */
+	pipefd[0] = r_fd;
+	pipefd[1] = w_fd;
+
+	return ret;
+
+ERR_W_FD:
+	fdtab_put_fd(tab, r_fd);
+ERR_VFS_INSTALL:
+	pipe_file_free(pipe_file);
+ERR_EXIT:
+	UK_ASSERT(ret < 0);
+	return ret;
+}
+
+/* TODO find a more efficient way to implement pipe2() */
+UK_SYSCALL_R_DEFINE(int, pipe2, int*, pipefd, int, flags)
+{
+	int rc;
+
+	rc = uk_syscall_r_pipe(pipefd);
+	if (rc)
+		return rc;
+
+	if (flags & O_CLOEXEC) {
+		fcntl(pipefd[0], F_SETFD, FD_CLOEXEC);
+		fcntl(pipefd[1], F_SETFD, FD_CLOEXEC);
+	}
+	if (flags & O_NONBLOCK) {
+		fcntl(pipefd[0], F_SETFL, O_NONBLOCK);
+		fcntl(pipefd[1], F_SETFL, O_NONBLOCK);
+	}
+	return 0;
+}
