+++ lib/vfscore/Makefile.uk
@@ -52,7 +52,6 @@ UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += stat-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += mkdir-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += mkdirat-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += umount2-2
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += pipe2-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += symlink-2
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += unlink-1
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += chroot-1
@@ -61,5 +60,4 @@ UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += getdents64-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += newfstatat-4
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += open-3
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += openat-4
-UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += pipe-1
 UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += creat-2
