+++ lib/vfscore/exportsyms.uk
@@ -169,12 +169,6 @@ vn_settimes
 vn_stat
 vn_unlock
 vfs_busy
-pipe
-uk_syscall_e_pipe
-uk_syscall_r_pipe
-pipe2
-uk_syscall_e_pipe2
-uk_syscall_r_pipe2
 mkfifo
 futimes
 uk_syscall_e_futimesat
